Moved test that requires libtiff
